Add table for rose colours

+++ choco_roses/crm/views.py
@@ -34,17 +34,7 @@ def orders_req(request):
 	search_value = request.GET.get('search_order', None)
 	search_status = request.GET.get('search_status', None)
 
-
-	if search_value and search_status != 'Все':
-		orders = search_order(search_value, search_status)
-	elif search_value and search_status == 'Все':
-		orders = search_order(search_value, search_status)
-	elif search_status and search_status != 'Все':
-		orders = Order.objects.filter(order_status=search_status)
-	else:
-		orders = Order.objects.all().order_by('-number')
-
-
+	orders = search_order(search_value, search_status)
 	paginator = Paginator(orders, 10)
 	page_number = request.GET.get('page', 1)
 	page_obj = paginator.get_page(page_number)
@@ -165,9 +155,11 @@ def rose_boxes(request):
 
 @staff_member_required
 def add_box(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	box_name = request.POST.get('box_name')
 	try:
-		box = RoseBoxes.objects.create(rose_box=box_name)
+		box = RoseBoxes.objects.create(box=box_name, status=True)
 	except TypeError as exc:
 		return JsonResponse({'error': f'{exc}'}, status=500)
 	except Exception as exc:
@@ -178,6 +170,8 @@ def add_box(request):
 
 @staff_member_required
 def status_box(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	box_status = request.POST.get('box_status')
 	box_id = request.POST.get('box_id')
 	try:
@@ -196,6 +190,8 @@ def status_box(request):
 
 @staff_member_required
 def delete_box(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	box_id = request.POST.get('box_id')
 	try:
 		RoseBoxes.objects.get(id=box_id).delete()
@@ -211,11 +207,13 @@ def delete_box(request):
 
 @staff_member_required
 def update_box(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	box_name = request.POST.get('box_name')
 	box_id = request.POST.get('box_id')
 	try:
 		box = RoseBoxes.objects.get(id=box_id)
-		box.rose_box = box_name
+		box.box = box_name
 		box.save()
 	except RoseBoxes.DoesNotExist as exc:
 		return JsonResponse({'error': f'{exc}'}, status=500)
@@ -229,6 +227,7 @@ def update_box(request):
 
 # endregion
 
+# region packings
 
 @staff_member_required
 def rose_packings(request):
@@ -245,9 +244,11 @@ def rose_packings(request):
 
 @staff_member_required
 def add_packing(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	packing_name = request.POST.get('packing_name')
 	try:
-		packing = RosePacking.objects.create(rose_packing=packing_name)
+		packing = RosePacking.objects.create(packing=packing_name, status=True)
 	except Exception as exc:
 		return JsonResponse({'error': f'{exc}'}, status=500)
 
@@ -256,6 +257,8 @@ def add_packing(request):
 
 @staff_member_required
 def status_packing(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	packing_status = request.POST.get('packing_status')
 	packing_id = request.POST.get('packing_id')
 	try:
@@ -274,36 +277,102 @@ def status_packing(request):
 
 @staff_member_required
 def delete_packing(request):
-	packing_id = request.POST.get('packing_id')
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	try:
+		packing_id = request.POST.get('packing_id')
 		RosePacking.objects.get(id=packing_id).delete()
 	except Exception as exc:
-		return JsonResponse({'error': f'{exc}'}, status=500)
+		return JsonResponse({'error': exc}, status=500)
 
 	return JsonResponse({'response': 'good'}, status=200)
 
 
 @staff_member_required
 def update_packing(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
 	packing_id = request.POST.get('packing_id')
 	packing_name = request.POST.get('packing_name')
 	try:
 		packing = RosePacking.objects.get(id=packing_id)
-		packing.rose_packing = packing_name
+		packing.packing = packing_name
 		packing.save()
 	except Exception as exc:
 		return JsonResponse({'error': exc}, status=500)
 	return JsonResponse({'response': 'good'})
 
 
-@staff_member_required
-def add_rose_colour(request):
-	pass
+# endregion
 
 
 @staff_member_required
-def delete_rose_colour(request):
-	pass
+def rose_colours(request):
+	colours = RoseColour.objects.all()
+	context = {'colours': colours}
+	return render(request, 'crm/colours.html', context=context)
+
+
+@staff_member_required
+def add_colour(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
+	try:
+		colour = RoseColour.objects.create(colour=request.POST.get('colour'), status=True)
+	except Exception as exc:
+		return JsonResponse({'error': exc}, status=500)
+
+	return JsonResponse({'response': 'good', 'colour_id': colour.id}, status=200)
+
+
+@staff_member_required
+def delete_colour(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
+	try:
+		colour_id = request.POST.get('colour_id')
+		RoseColour.objects.get(id=colour_id).delete()
+	except Exception as exc:
+		return JsonResponse({'error': exc}, status=500)
+	return JsonResponse({'response': 'good'})
+
+
+@staff_member_required
+def update_colour(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
+	colour_id = request.POST.get('colour_id')
+	colour = request.POST.get('colour')
+	print(colour)
+	try:
+		colour_obj = RoseColour.objects.get(id=colour_id)
+		colour_obj.colour = colour
+		colour_obj.save()
+	except TypeError as exc:
+		return JsonResponse({'error': f'{exc}'}, status=500)
+	except Exception as exc:
+		return JsonResponse({'error': f'{exc}'}, status=500)
+	return JsonResponse({'response': 'good'})
+
+
+@staff_member_required
+def status_colour(request):
+	if not is_ajax(request):
+		return JsonResponse({'error': 'request in not ajax'}, status=400)
+	colour_status = request.POST.get('colour_status')
+	colour_id = request.POST.get('colour_id')
+	try:
+		colour = RoseColour.objects.get(id=colour_id)
+		colour.status = colour_status
+		colour.save()
+	except RoseColour.DoesNotExist as exc:
+		return JsonResponse({'error': f'{exc}'}, status=500)
+	except TypeError as exc:
+		return JsonResponse({'error': f'{exc}'}, status=500)
+	except Exception as exc:
+		return JsonResponse({'error': f'{exc}'}, status=500)
+
+	return JsonResponse({'response': 'good'})
 
 
 @staff_member_required
